Change location of where styles are defined
Floated the image on the right which has caused the title to be centered but moved the rest of the heading off centre. To fix that and also to enhance visually, removed center align from body and moved it to h1 only. Also modified some of the list items so they no longer have bullet points. For recipes where the description has multiple sub-headings, created a new class so they do not have bullet points and have bold text.